requirements.txt
streamlit>=1.50.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pypdf>=3.17.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0

